requirements.txt
flask==2.2.5
werkzeug==2.2.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
retell-sdk==4.30.0
jinja2==3.1.2
pathlib==1.0.1 